ui(settings): refactor settings into sidebar-navigated panel with premium controls (#81)
- Split 888-line monolith into sidebar + 4 sections (General, System, Gateways, About)
- Add sidebar nav with animated indicator (Framer Motion layoutId)
- Replace green segmented controls with neutral pill-style SegmentedControl
- Add auto theme mode (follows system via prefers-color-scheme)
- Add gateway type field (openclaw) with badge on cards
- Add delete confirmation dialog for gateway removal
- Fix edit form positioning: renders inline below the edited card
- Redesign gateway card layout with status column and action separator
