+++ packages/desktop/src/renderer/layouts/Settings/components/SegmentedControl.tsx
@@ -1,0 +1,49 @@
+import type { ReactNode } from 'react';
+import { motion } from 'framer-motion';
+import { cn } from '@/lib/utils';
+
+interface SegmentedControlProps<T extends string> {
+  value: T;
+  onChange: (value: T) => void;
+  options: { value: T; label: ReactNode }[];
+  layoutId: string;
+  ariaLabel?: string;
+}
+
+export default function SegmentedControl<T extends string>({
+  value,
+  onChange,
+  options,
+  layoutId,
+  ariaLabel,
+}: SegmentedControlProps<T>) {
+  return (
+    <div role="radiogroup" aria-label={ariaLabel} className="relative flex rounded-lg bg-[var(--bg-tertiary)] p-0.5">
+      {options.map((opt) => (
+        <button
+          key={opt.value}
+          type="button"
+          role="radio"
+          aria-checked={value === opt.value}
+          onClick={() => onChange(opt.value)}
+          className={cn(
+            'relative flex items-center gap-1.5 px-3.5 py-1.5 text-sm transition-colors',
+            'focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-[var(--ring-accent)]',
+            value === opt.value
+              ? 'text-[var(--text-primary)]'
+              : 'text-[var(--text-secondary)] hover:text-[var(--text-primary)]',
+          )}
+        >
+          {value === opt.value && (
+            <motion.div
+              layoutId={layoutId}
+              className="absolute inset-0 rounded-md bg-[var(--bg-elevated)] shadow-sm border border-[var(--border-subtle)]"
+              transition={{ type: 'spring', bounce: 0.15, duration: 0.4 }}
+            />
+          )}
+          <span className="relative z-10 flex items-center gap-1.5">{opt.label}</span>
+        </button>
+      ))}
+    </div>
+  );
+}
